refactor(submission): resonse -> submission으로 도메인명 변경
