--- project/sungwoong-onboarding/src/main/java/com/multi/sungwoongonboarding/responses/application/repository/AnswerRepository.java
@@ -1,14 +1,0 @@
-package com.multi.sungwoongonboarding.responses.application.repository;
-
-import com.multi.sungwoongonboarding.responses.domain.Answers;
-
-import java.util.List;
-
-public interface AnswerRepository {
-
-    Answers save(Answers answers);
-
-    List<Answers> saveAll(List<Answers> answers);
-
-    List<Answers> findAll();
-}
