--- project/sungwoong-onboarding/src/main/java/com/multi/sungwoongonboarding/responses/infrastructure/answers/AnswersJpaRepository.java
@@ -1,12 +1,0 @@
-package com.multi.sungwoongonboarding.responses.infrastructure.answers;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface AnswersJpaRepository extends JpaRepository<AnswerJpaEntity, Long> {
-    
-    List<AnswerJpaEntity> findByResponseJpaEntityId(Long responseId);
-    
-    List<AnswerJpaEntity> findByQuestionId(Long questionId);
-}
