--- project/sungwoong-onboarding/src/main/java/com/multi/sungwoongonboarding/responses/infrastructure/responses/ResponsesJpaRepository.java
@@ -1,12 +1,0 @@
-package com.multi.sungwoongonboarding.responses.infrastructure.responses;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface ResponsesJpaRepository extends JpaRepository<ResponseJpaEntity, Long> {
-    
-    List<ResponseJpaEntity> findByFormId(Long formId);
-    
-    List<ResponseJpaEntity> findByUserId(String userId);
-}
