+++ project/sungwoong-onboarding/src/main/java/com/multi/sungwoongonboarding/submission/application/repository/AnswerRepository.java
@@ -1,0 +1,14 @@
+package com.multi.sungwoongonboarding.submission.application.repository;
+
+import com.multi.sungwoongonboarding.submission.domain.Answers;
+
+import java.util.List;
+
+public interface AnswerRepository {
+
+    Answers save(Answers answers);
+
+    List<Answers> saveAll(List<Answers> answers);
+
+    List<Answers> findAll();
+}
