+++ project/sungwoong-onboarding/src/main/java/com/multi/sungwoongonboarding/submission/application/repository/ResponseRepository.java
@@ -1,0 +1,14 @@
+package com.multi.sungwoongonboarding.submission.application.repository;
+
+import com.multi.sungwoongonboarding.submission.domain.Responses;
+
+import java.util.List;
+
+public interface ResponseRepository {
+
+    Responses save(Responses responses);
+
+    List<Responses> findAll();
+
+    Responses findById(Long id);
+}
