+++ project/sungwoong-onboarding/src/main/java/com/multi/sungwoongonboarding/submission/dto/AnswerResponse.java
@@ -1,0 +1,20 @@
+package com.multi.sungwoongonboarding.submission.dto;
+
+import com.multi.sungwoongonboarding.submission.domain.Answers;
+import lombok.Builder;
+import lombok.RequiredArgsConstructor;
+
+@Builder
+@RequiredArgsConstructor
+public class AnswerResponse {
+
+    private final Long id;
+    private final Long questionId;
+    private final Long optionId;
+    private final String answerText;
+
+    public static AnswerResponse fromDomain(Answers answers) {
+        return new AnswerResponse(answers.getId(), answers.getQuestionId(), answers.getOptionId(), answers.getAnswerText());
+    }
+
+}
