+++ project/sungwoong-onboarding/src/main/java/com/multi/sungwoongonboarding/submission/infrastructure/answers/AnswersJpaRepository.java
@@ -1,0 +1,12 @@
+package com.multi.sungwoongonboarding.submission.infrastructure.answers;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface AnswersJpaRepository extends JpaRepository<AnswerJpaEntity, Long> {
+    
+    List<AnswerJpaEntity> findByResponseJpaEntityId(Long responseId);
+    
+    List<AnswerJpaEntity> findByQuestionId(Long questionId);
+}
